+++ spec/Suite/AllowSpec.php
@@ -23,7 +23,7 @@ describe("Allow", function() {
     /**
      * Save current & reinitialize the Interceptor class.
      */
-    before(function() {
+    beforeAll(function() {
         $this->previous = Interceptor::instance();
         Interceptor::unpatch();
 
@@ -37,7 +37,7 @@ describe("Allow", function() {
     /**
      * Restore Interceptor class.
      */
-    after(function() {
+    afterAll(function() {
         Interceptor::load($this->previous);
     });