+++ spec/Suite/Cli/KahlanSpec.php
@@ -14,7 +14,7 @@ describe("Kahlan", function() {
     /**
      * Save current & reinitialize the Interceptor class.
      */
-    before(function() {
+    beforeAll(function() {
         $this->previous = Interceptor::instance();
         Interceptor::unpatch();
     });
@@ -22,7 +22,7 @@ describe("Kahlan", function() {
     /**
      * Restore Interceptor class.
      */
-    after(function() {
+    afterAll(function() {
         Interceptor::load($this->previous);
     });