+++ spec/Suite/Matcher/ToBeCalledSpec.php
@@ -17,7 +17,7 @@ describe("ToBeCalled", function() {
         /**
          * Save current & reinitialize the Interceptor class.
          */
-        before(function() {
+        beforeAll(function() {
             $this->previous = Interceptor::instance();
             Interceptor::unpatch();
 
@@ -31,7 +31,7 @@ describe("ToBeCalled", function() {
         /**
          * Restore Interceptor class.
          */
-        after(function() {
+        afterAll(function() {
             Interceptor::load($this->previous);
         });