+++ spec/Suite/Plugin/StubSpec.php
@@ -22,7 +22,7 @@ describe("Stub", function() {
     /**
      * Save current & reinitialize the Interceptor class.
      */
-    before(function() {
+    beforeAll(function() {
         $this->previous = Interceptor::instance();
         Interceptor::unpatch();
 
@@ -36,7 +36,7 @@ describe("Stub", function() {
     /**
      * Restore Interceptor class.
      */
-    after(function() {
+    afterAll(function() {
         Interceptor::load($this->previous);
     });